Fix interrupting ModelCheckingJob and LTLCheckingJob
When interrupting while a checking job is running, the step command
can get interrupted instead. Previously this was not detected, and the
job would continue running, because it was not completed or interrupted
(only the step command was). This fixes the issue by setting the job
command as interrupted if the step command is interrupted.