Fix null handling in ProBError
The errorItems list may be null, and it means something different than
an empty list. An empty list means that the CLI returned an empty list
of errors, but null means that no error list is available (for example
when a ProBError is thrown manually from Java).
This may not be the best design though, perhaps it would be better to
use an empty list in both cases? But that would mean that the
"ProB returned no error messages" message would be added even in places
where no error messages can possibly be returned (because the CLI
is not directly involved).