Fix GetMachineStructureCommand for machines with animation function
The pretty print of animation function images is the image index, as an
integer rather than an atom. This previously caused an exception,
because PrologTerm.atomicString does not accept integers.
I don't know if integers should be allowed as pretty prints, or if this
should be changed on the Prolog side to convert the integer to an atom.