Improve exception thrown when terms are sent during CLI shutdown
The term is now only written to the log instead of added to the
exception message, to avoid displaying very long messages to the user
(for example when the ProB 2 UI displays the exception).
The exception type is now IOException, so the method throws one
exception type consistently in case of CLI communication failures.