Fix Groovy tests hanging on fail
Some tests create ModelCheckers, which create thread pool executors.
The executors need to be shut down to stop the pool threads, but
ModelChecker does not expose a way to shut down its executor.
Instead we need to make sure that no ModelCheckers are stored in
global variables in the test scripts, so that the garbage collector
will collect them (which shuts down the executors as well, because
they have a finalizer). For some reason, global variables in the test
scripts are not garbage-collected.