Change Groovy tests to use script-local variables instead of globals
This allows better analysis of the test sources with e. g. IntelliJ,
because now the types of (almost) all variables can be inferred.
This can be used to find class/method usages properly, instead of
having to use a plain text search.

Previously this was not possible, because all objects were stored in
global properties. IntelliJ is not able to infer the types for these,
because they are not declared anywhere and could theoretically have
Groovy magic behind them. (Of course that is not the case here, but
IntelliJ has no way to know that.)