Fix setting cspmf as executable in Installer
Previously *only* the user executable bit was set, and all other
permission bits were unset. Now all executable bits are set and
existing ones are kept, as expected.
The installer also re-adds the owner read and write bits on the cspmf
executable if are missing. (This is a temporary fix and can be removed
again after a while.)