Move Tuple2/ArrayList equality check from Tuple2 to initscript
The previous implementation of Tuple2.equals did not follow the
contract of equals, because it was not commutative (tuple.equals(list)
could be true, but list.equals(tuple) was not).

The equality check is now implemented in both directions in the
initscript. However, this means that code outside of the ProB 2 Groovy
environment now needs to convert explicitly between tuples and lists
when doing equality checks.