Set Logback config file explicitly instead of using the default name
This avoids conflicts with other code that also uses the default
logback.xml. It also removes the need to have a single top-level
config that just includes another config - now we can switch between
multiple configs by changing the path in the property.

This change does not apply to the logback-test.xml - since it is only
used when running the tests, there is no risk of conflicts there.